requirements.txt
transformers==4.34.0
torch==2.0.0
pillow==9.0.0
pandas==1.5.0
tqdm==4.64.1
